requirements.txt
Django==3.2.13
djangorestframework==3.11.0
django-environ==0.9.0
django-cors-headers==3.10.1
django-allauth==0.42.0
pytest
